refactor: move expression pkg to internal
