+++ tools/ghx/actions/context.go
@@ -10,7 +10,7 @@ import (
 	"github.com/aweris/gale/internal/fs"
 
 	"github.com/aweris/gale/internal/core"
-	"github.com/aweris/gale/tools/ghx/expression"
+	"github.com/aweris/gale/internal/expression"
 )
 
 var _ expression.VariableProvider = new(ExprContext)
