+++ tools/ghx/runner/cmd_executor.go
@@ -10,9 +10,9 @@ import (
 	"strings"
 
 	"github.com/aweris/gale/internal/core"
+	"github.com/aweris/gale/internal/expression"
 	"github.com/aweris/gale/internal/log"
 	"github.com/aweris/gale/tools/ghx/actions"
-	"github.com/aweris/gale/tools/ghx/expression"
 )
 
 var _ Executor = new(CmdExecutor)
